requirements.txt
typer>=0.9.0
questionary>=2.0.0
pyyaml>=6.0
rich>=13.0.0
requests>=2.31.0
pydantic>=2.0.0
